requirements.txt
# requirements.txt
# Dependencies for the HDAB-NL pseudonymization prototype

cryptography==43.0.3
python-dateutil>=2.8.2
liboqs-python>=0.12.0